mcp-servers/content-automation/pyproject.toml
[project]
name = "content-automation"
version = "0.1.0"
description = "블로그 + 카드뉴스 자동 생성 MCP"
requires-python = ">=3.10"
dependencies = [
    "mcp>=0.9.0",
    "anthropic>=0.39.0",
    "python-dotenv>=1.0.0"
]

[project.scripts]
content-automation = "server:main"